Implemented a listening mechanism for reloads of scoped beans and implemented clearing the session of beans whose class has been reloaded.
Prior to this, a ClassCastException would occur when retrieving a session scoped bean from the session as it's old class was different to what the new proxy was expecting.

The mechanism is extensible in that users can register their own ScopedBeanReloadListener in the application context to be notified of reloads.